pyproject.toml
[project]
name = "boggleywollah"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.37.9",
    "deltalake>=0.25.4",
    "getdaft>=0.4.7",
]
